image Stitching/requirements.txt
flask>=3.0
opencv-python>=4.8
numpy>=1.24
torch
torchvision
timm>=0.9
Pillow>=10.0
